+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Upgrade react-dom@^16.3.0
 - Upgrade styled-components@^4.1.3
 - Remove built-in Header in latest-section of index-page
+- Remove text-decoration of links in Footer
 
 ### 5.1.3
 - Add url event label to google analysis click event